rename encyption/sha1 to crypto/sha1
