+++ components/crypto/sha1/index.ts
