+++ components/crypto/sha1/sha1.ts
