+++ scopes/preview/preview/pre-bundle-utils.ts
@@ -4,7 +4,7 @@ import { UIRoot } from '@teambit/ui';
 import { getAspectDirFromBvm } from '@teambit/aspect-loader';
 import { SlotRegistry } from '@teambit/harmony';
 import { ArtifactDefinition } from '@teambit/builder';
-import { sha1 } from '@teambit/toolbox.encryption.sha1';
+import { sha1 } from '@teambit/toolbox.crypto.sha1';
 
 /// utils
