+++ scopes/workspace/workspace-config-files/writers/extending-config-files.ts
@@ -1,6 +1,6 @@
 import normalize from 'normalize-path';
 import format from 'string-format';
-import { sha1 } from '@teambit/toolbox.encryption.sha1';
+import { sha1 } from '@teambit/toolbox.crypto.sha1';
 import fs from 'fs-extra';
 import { dirname, join, relative } from 'path';
 import pMapSeries from 'p-map-series';
